create a separate class for hash-based IDs
