+++ src/ID.ts
@@ -1,6 +1,4 @@
-import crypto from "node:crypto";
-
-export default class ID {
+export default abstract class ID {
     readonly #id: string;
 
     /**
@@ -17,25 +15,17 @@ export default class ID {
         return this.#id;
     }
 
+    /**
+     * Get string representation of this ID
+     */
     public toString(): string {
         return this.id;
     }
 
+    /**
+     * Check if another ID is for the same resource as this one
+     */
     public equals(other: ID): boolean {
         return this instanceof other.constructor && this.id === other.id;
-    }
-
-    private static hash(string: string): string {
-        return crypto.createHash("sha1").update(string).digest("base64").replace(/=+$/, "").replace(/\/+/g, "-");
-    }
-
-    /**
-     * Calculate ID from strings/IDs
-     */
-    protected static of(...args: (string | ID | null | undefined)[]): ID {
-        return new this(ID.hash(args.map(s => {
-            const k = (typeof s === "string") ? s : s?.id ?? "";
-            return k.length + k;
-        }).join("")));
     }
 }
